Apiroute (#42)
* added more and more functions


* fixed iterator


* filter in data fetch added


* added query in getServerSideProps


* added detailed analysis


* router.push replaced with next link component


* updated styles


---------
